+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 <small>
 
 * Version: 0.12.0
-* Last update: 230410
+* Last update: 230411
 </small>
 
 ***
@@ -40,6 +40,10 @@
     * [ ] revenue
     * [ ] solution
 
+* wip... (B) 230411piu_a fix OPISS.008 - bug @ ARE main page loader, no ADS data in a business domain
+    * [x] step 1 - make a solve plan in file `/are/are_main.py` about line #
+    * [ ] step 2 - solve bug for first bss domain, as order in code, just for test
+    * [ ] step 3 - generalize for all bss domains
 * 230410piu_a ARE start page redirect to Sales Project choose in case sent PK does not exists (location: `are_main.py`)
 * 230409piu_a update chart generator function `are_chart(business_domain, chart_type)` (module `are/are_main.py`, route: `'/are_chart/<business_domain>/<chart_type>'`) to generate unique image files with auto removing after usage